Refactor large portions of code, change way grade data is stored
I refactored large portions of the code dealing with storing grade data. Instead of storing data as whatever pickle spits out, it now stores it all in human-readable JSON. In addition, it now keeps track of the grades by the Canvas ID of the course instead of the user-given nickname of the course, making it support changing the course nickname without messing things up.
It also splits up storage of the grades and the course data, only updating course data when it changes, which keeps it from storing too much redundant data in the score-storage files, and allows for continued use of a remembered color if the color/course isn't fetched on a particular cycle.